Fix loading model from jar on Java 11
When loading a model from a jar file, Zoltar's copyDir method copies the model to a
temporary directory. copyDir calculates the path of the file being copied relative
to the source root directory. It does this using String.substring, and assumes the
source root directory has a trailing slash. In Java 11, ZipPath (which represents
Paths in jar files) removes the trailing slash. This throws off substring selection,
and causes Zoltar to try to copy files to "/", which fails with AccessDeniedException.

This PR replaces the substring operation with Path.relativize which works correctly
whether the trailing slash is present or not.