Add trailing slash only for GCS paths
When copying files from a jar file, we cannot have a trailing slash on the src
URI due to quirks in ZipPath.relativize in Java 8 (not a problem in Java 11).
"/foo/".relativize("/foo/bar") returns "../bar" instead of "bar", which causes us
to build an invalid destination directory.

But when copying from GCS, we must have a trailing slash. With this commit, we will
add the trailing slash for GCS paths only.