Fix for TensorFlowGraphLoader failure when loading model from GCS (#180)
PR #176 made it so that a trailing slash was appended to GCS URIs. This was
to fix downloading tensorflow directories, which requires the trailing slash.
This broke TensorFlowGraphLoader, because that downloads a single file, and
the GCS client throws an exception if a trailing slash is used with a file.

This PR makes it so that the trailing slash is appended by the TensorFlowLoader
class, not FileSystemExtras. The loader classes know if they are downloading
a file or a directory (FileSystemExtras does not), so they need to be
responsible for handling the slash.

Also added LoaderIT to test TensorFlowGraphLoader and TensorFlowLoader under
various conditions.